backend/requirements.txt
Django==5.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.4.0
django-filter==24.3
PyMySQL==1.1.2
python-dotenv==1.0.1
